requirements.txt
pandas
numpy
plotly
matplotlib
pillow